--- src/runtime/config.ts
@@ -1,2 +1,0 @@
-export const moduleName = 'nuxt-content-assets'
-export const moduleKey = 'contentAssets'
