+++ src/runtime/utils/index.ts
@@ -5,3 +5,4 @@ export * from './build'
 export * from './fs'
 export * from './string'
 export * from './object'
+export { makeIgnores } from './config'
